app/requirements.txt
vosk
pyaudio
RPi.GPIO
reportlab
ctransformers
Flask